rename impactView to index
